Improved Gateway Execution, Error Handling and Link management.
CrtpLinks can now be closed on on crazyradio #closes 7
The add crazyflie service now responds with the error message associated #closes 10
Crazyflie now gets transitioned from unconfigured with from the gateway.
If adding to tracked or broadcaster raises an error this will be in the service.
This is a major commit. Test rigorously before merging with main.